requirements.txt
pycrypto
Django==1.10.2
django-notifications-hq==1.2
django-multiselectfield
djangorestframework
